Extract admin header to own file
