+++ client/src/pages/admin/index.tsx
@@ -8,27 +8,22 @@ import Loading from '../../components/Loading';
 import { postRequest } from '../../api';
 import { errorAlert } from '../../util';
 import { useNavigate } from 'react-router-dom';
-import Button from '../../components/Button';
 import { useAdminStore, useClockStore, useFlagsStore, useOptionsStore } from '../../store';
-import PauseButton from '../../components/admin/PauseButton';
-import ConfirmPopup from '../../components/ConfirmPopup';
 import ToTopButton from '../../components/ToTopButton';
+import AdminHeader from '../../components/admin/AdminHeader';
 
 const Admin = () => {
     const navigate = useNavigate();
     const fetchStats = useAdminStore((state) => state.fetchStats);
-    const clock = useClockStore((state) => state.clock);
     const fetchClock = useClockStore((state) => state.fetchClock);
     const fetchProjects = useAdminStore((state) => state.fetchProjects);
     const fetchJudges = useAdminStore((state) => state.fetchJudges);
-    const options = useOptionsStore((state) => state.options);
     const fetchOptions = useOptionsStore((state) => state.fetchOptions);
     const fetchFlags = useFlagsStore((state) => state.fetchFlags);
 
     const [showProjects, setShowProjects] = useState(true);
     const [loading, setLoading] = useState(true);
     const [lastUpdate, setLastUpdate] = useState(new Date());
-    const [swapPopup, setSwapPopup] = useState(false);
 
     useEffect(() => {
         // Check if user logged in
@@ -68,81 +63,17 @@ const Admin = () => {
         return () => clearInterval(refresh);
     }, []);
 
-    const swapJudgeGroups = async () => {
-        const res = await postRequest<OkResponse>('/admin/groups/swap', 'admin', null);
-        if (res.status !== 200) {
-            errorAlert(res);
-            return;
-        }
-        setSwapPopup(false);
-        fetchJudges();
-        fetchOptions();
-        alert('Judge groups swapped successfully');
-    };
-
     if (loading) {
         return <Loading disabled={!loading} />;
     }
     return (
         <>
             <JuryHeader withLogout isAdmin />
-            <div className="absolute top-4 w-full flex items-center z-0">
-                <PauseButton />
-                <Button
-                    type="outline"
-                    onClick={() => {
-                        navigate('/admin/settings');
-                    }}
-                    small
-                    bold
-                    className="ml-4 py-2"
-                >
-                    Settings
-                </Button>
-                <Button
-                    type="outline"
-                    onClick={() => navigate('/admin/log')}
-                    small
-                    bold
-                    className="ml-4 py-2"
-                >
-                    Audit Log
-                </Button>
-                {options && options.multi_group && options.switching_mode === 'manual' && (
-                    <>
-                        <p className="text-lg text-right grow mr-4 text-lighter">
-                            <span className="font-bold">Swaps: </span>
-                            {options.manual_switches}
-                        </p>
-                        <Button
-                            type="gold"
-                            onClick={setSwapPopup.bind(null, true)}
-                            disabled={clock.running}
-                            tooltip="Groups can only be swapped when judging is paused"
-                            small
-                            bold
-                            className="mr-36 py-2"
-                        >
-                            Swap Judge Groups
-                        </Button>
-                    </>
-                )}
-            </div>
+            <AdminHeader />
             <AdminStatsPanel />
             <AdminToggleSwitch state={showProjects} setState={setShowProjects} />
             <AdminToolbar showProjects={showProjects} lastUpdate={lastUpdate} />
             <AdminTable showProjects={showProjects} />
-            <ConfirmPopup
-                enabled={swapPopup}
-                setEnabled={setSwapPopup}
-                onSubmit={swapJudgeGroups}
-                submitText="Swap"
-                title="Swap Judge Groups"
-                red
-            >
-                Are you sure you want to swap judge groups? This will increment the group number of
-                every judge.
-            </ConfirmPopup>
             <ToTopButton />
         </>
     );
